move helper into `testutils`
