+++ tests/sentry/utils/test_sdk.py
@@ -1,4 +1,3 @@
-from typing import Any, Callable
 from unittest.mock import MagicMock, patch
 
 from django.http import HttpRequest
@@ -6,29 +5,13 @@ from rest_framework.request import Request
 from sentry_sdk import Scope
 
 from sentry.testutils import TestCase
+from sentry.testutils.helpers import set_mock_context_manager_return_value
 from sentry.utils.sdk import (
     capture_exception_with_scope_check,
     check_current_scope_transaction,
     check_tag,
     merge_context_into_scope,
 )
-
-
-# TODO: This is kind of gross, but no other way seemed to work
-def set_mock_context_manager_return_value(
-    context_manager_constructor: Callable, as_value: Any
-) -> None:
-    """
-    Ensure that if the code being tested includes something like
-
-        with some_func() as some_value:
-
-    then `some_value` will equal `as_value`.
-
-    Note: `context_manager_constructor` should be `some_func`, not `some_func()`.
-    """
-
-    context_manager_constructor.return_value.__enter__.return_value = as_value
 
 
 class SDKUtilsTest(TestCase):
